--- towhee/data_source/__init__.py
@@ -1,34 +1,0 @@
-# Copyright 2021 Zilliz. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-import abc
-from collections.abc import Iterable
-
-
-class DataSource(Iterable):
-    """Base class of data source
-    """
-
-
-class InputData(metaclass=abc.ABCMeta):
-    """Base class of input data
-    """
-    @abc.abstractmethod
-    def get_data(self) -> dict:
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def on_finish(self) -> bool:
-        raise NotImplementedError
