+++ towhee/dataframe/array/__init__.py
@@ -12,9 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from towhee.array.array import Array
-from towhee.array.utils import full
-
+from towhee.dataframe.array.array import Array
+from towhee.dataframe.array.utils import full
 
 __all__ = [
     'Array',
